+++ DOCS/CHANGELOG.md
@@ -34,7 +34,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 - new traits: `ExtCellOption`, `ExtOptRes`, `MemPod`.
 - new types: `UnitBi`, `UnitSi` and trait: `Unit`.
 - new types: `CompressionMode`, `EncodingMode`.
-- new items: `RendError`, `RendResult`, `Pnm`.
+- new items: `Pnm`, `RendError`, `RendResult`, `ImageError`, `ImageResult`.
 - new type: `Xoroshiro128pp`.
 - new type: `Interval`.
 - new type: `FatPtr`.